Moved IndexSortKey to metadata package
